moved types for libs aside from std to own stubs
