+++ python/ry/ryo3/_brotli.pyi
@@ -1,0 +1,13 @@
+"""ryo3-brotli types"""
+
+from __future__ import annotations
+
+# =============================================================================
+# BROTLI
+# =============================================================================
+def brotli_encode(
+    input: bytes, quality: int = 11, magic_number: bool = False
+) -> bytes: ...
+def brotli_decode(input: bytes) -> bytes: ...
+def brotli(input: bytes, quality: int = 11, magic_number: bool = False) -> bytes:
+    """Alias for brotli_encode"""
