+++ python/ry/ryo3/_same_file.pyi
@@ -1,0 +1,7 @@
+"""ryo3-same-file types"""
+
+from __future__ import annotations
+
+from os import PathLike
+
+def is_same_file(a: PathLike[str], b: PathLike[str]) -> bool: ...
