+++ python/ry/ryo3/_shlex.pyi
@@ -1,0 +1,7 @@
+"""ryo3-shlex types"""
+
+from __future__ import annotations
+
+def shplit(s: str) -> list[str]:
+    """shlex::split wrapper much like python's stdlib shlex.split but faster"""
+    ...
